Crime datasets files
For the files from 2021 and 2022, the one with the most information about Florida specifically is the one that says "Florida Offense type by Agency" and "city-data". The rest is data for the entire US in case we want to do comparisons with the rest of the country. Keep in mind most files are xlsx files, you need to download the csv file from excel for the file that you want to analyze with Python.